server side personal and public profile page
